MANO template mesh positioned at root and pre-rotated to face camera
