+++ src/ireg_layer.py
@@ -4,17 +4,20 @@ import tensorflow as tf
 
 # TODO: Make it so that we do not have to pass in batch_size like this.
 def MAKE_REGRESSION_MODULE(batch_size):
+  
+  MODEL_OUPUTS = 58
+  
   inputs = tf.keras.layers.Input(shape=[576,])
   
   model = tf.keras.models.Sequential()
-  model.add(tf.keras.Input(shape = (637,)))
+  model.add(tf.keras.Input(shape = (576 + MODEL_OUPUTS,)))
   model.add(tf.keras.layers.Dense(288, activation = 'relu'))
   model.add(tf.keras.layers.Dropout(0.5))
   model.add(tf.keras.layers.Dense(288, activation = 'relu'))
   model.add(tf.keras.layers.Dropout(0.5))
   model.add(tf.keras.layers.Dense(61, activation = 'relu'))
   
-  params = tf.zeros((batch_size,61))
+  params = tf.zeros((batch_size, MODEL_OUPUTS))
   model.summary()
 
   for i in range(3):
