+++ src/mano_layer.py
@@ -38,6 +38,28 @@ class MANO_Model(Model):
       self.J = tf.convert_to_tensor(self.J.todense(), dtype=tf.float32) # Need to convert sparse to dense matrix
       self.W = tf.convert_to_tensor(self.W, dtype=tf.float32)
 
+      # move T_Bar such that it is at (0,0,0)
+      root_pos = tf.constant([[[0.09566993,  0.00638343,  0.00618631]]])
+      self.T_bar -= root_pos
+
+      # Apply a rotation to make the MANO hand be in a more sensible default position.
+      batch_size = 1
+      beta = tf.zeros([batch_size, 10])
+      pose = tf.repeat(tf.constant([[
+          [ 
+            # NOTE: this rotation in axis-angle representation was 
+            # determined by the handy https://www.andre-gaschler.com/rotationconverter/
+            0.5773503 * 2.0943951, 
+            0.5773503 * 2.0943951, 
+            0.5773503 * 2.0943951
+          ], [0,0,0], [0,0,0], 
+          [0,0,0], [0,0,0], [0,0,0], [0,0,0], [0,0,0], [0,0,0],
+          [0,0,0], [0,0,0], [0,0,0], [0,0,0], [0,0,0], [0,0,0], [0,0,0]
+      ]], dtype=tf.float32), repeats=[batch_size], axis=0)
+      B_p = blend_pose(pose, self.P)
+      posed_joints, posed_mesh = lbs(pose, self.J, self.K, self.W, self.T_bar, B_p)
+      self.T_bar = posed_mesh[0] # Select just first batch.
+
       # indices are the RHD convention, the stored values are indices in MANO convention.
       self.remap_joints = tf.constant(
         [
@@ -104,7 +126,7 @@ class MANO_Model(Model):
     except Exception as e:
       print(e, "Unable to find MANO_RIGHT.pkl")
 
-  def call(self, beta, pose, root_trans, training=False):
+  def call(self, beta, pose, training=False):
 
     T_shaped = self.T_bar + blend_shape(beta, self.S)
     B_p = blend_pose(pose, self.P)
@@ -117,11 +139,11 @@ class MANO_Model(Model):
     posed_joints = tf.concat( [posed_joints, fingertips], axis=1)
     posed_joints = tf.gather(posed_joints, indices=self.remap_joints, axis=1)
 
-    root_trans1 = tf.expand_dims(root_trans, axis=1)
-    root_trans1 = tf.repeat(root_trans1,778, axis=1)
-    root_trans2 = tf.expand_dims(root_trans, axis=1)
-    root_trans2 = tf.repeat(root_trans2, 21, axis=1)
-    posed_mesh += root_trans1
-    posed_joints += root_trans2
+    # root_trans1 = tf.expand_dims(root_trans, axis=1)
+    # root_trans1 = tf.repeat(root_trans1,778, axis=1)
+    # root_trans2 = tf.expand_dims(root_trans, axis=1)
+    # root_trans2 = tf.repeat(root_trans2, 21, axis=1)
+    # posed_mesh += root_trans1
+    # posed_joints += root_trans2
  
     return posed_mesh, posed_joints
