+++ src/test_MANO_bp.py
@@ -6,18 +6,9 @@ import os
 import pickle
 import open3d as o3d
 import time
-from mano_bs_utest import blend_shape
+from mano_bs import blend_shape
 from rodrigues import rodrigues
-
-def blend_pose(pose, P):
-    batch_size = pose.shape[0]
-    I = tf.eye(3)
-    rotationMatrix = tf.reshape(rodrigues(tf.reshape(pose,(-1, 3))),(batch_size, -1, 3, 3))
-    pose_sum = tf.reshape(rotationMatrix[:, 1:, :, :]-I, (batch_size, -1))
-    blend_pose = tf.reshape(tf.linalg.matmul(pose_sum, tf.reshape(P, (135, -1))), (batch_size, -1, 3))
-    
-
-    return blend_pose
+from mano_bp import blend_pose
 
 if __name__ == "__main__":
