+++ src/test_MANO_bs.py
